move Void to model package
