+++ common/src/main/java/com/distkv/drpc/model/Void.java
@@ -1,4 +1,4 @@
-package com.distkv.drpc.common;
+package com.distkv.drpc.model;
 
 
 public class Void {
