+++ protocol/src/main/java/com/distkv/drpc/ServerImpl.java
@@ -7,7 +7,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import com.distkv.drpc.api.async.DefaultResponse;
 import com.distkv.drpc.api.async.Request;
 import com.distkv.drpc.api.async.Response;
-import com.distkv.drpc.common.Void;
+import com.distkv.drpc.model.Void;
 import com.distkv.drpc.exception.DrpcException;
 import com.distkv.drpc.utils.ReflectUtils;
